DEPORTAL/deidentification/requirements.txt
Django==4.2.17
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
jsonpickle==4.0.0
# notebook
jupyterlab==4.3.2
psycopg2==2.9.10
#psycopg2-binary
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-keycloak==5.1.1
gunicorn==23.0.0
dateparser
rapidfuzz
presidio_analyzer
presidio_anonymizer
snowflake-connector-python==3.12.4
snowflake-sqlalchemy==1.7.3
google-cloud-storage
pyodbc
pymssql
mysql-connector-python
pycryptodome
lxml
pyahocorasick
redis
# for the testing purpose to create dummy data
# faker

# added to fix docker build issue
#spacy==3.5.0
#blis==0.7.9; platform_system == "Linux" and platform_machine == "x86_64"
#numpy==1.24.0
#pandas
pynetdicom>=1.5.7,<2.0
pydicom
networkx
channels>=4.0
channels-redis>=4.2
daphne>=4.1
pandas
spacy==3.8.7

#phi-analyzer
aiohttp
google-generativeai
langchain
langchain-core
langgraph
ollama
openai